Add workerPlatform attribute to identify OS
Added a new attribute `workerPlatform` to classes related to tasks, such as `TaskNode`, `TaskInstance`, `ProcessInstance`, and `TaskDefinition`. This is crucial to differentiate OS types such as Unix, Windows, etc. for specific tasks. Additionally, updated `OSUtils` to specify OS types using an Enum and enhanced the task dispatching logic to consider workerPlatform when selecting a host for task execution.